apk build ready and bunch of thing done

+++ Assets/Scripts/EnemyStatus.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class EnemyStatus : MonoBehaviour
 {
-    public float Worth;
+    public int Worth;
     public float MoveMod;
     
 }
